Add support for HAProxy LOCAL header
While the PROXY header is used for proxied connections, the LOCAL header
is used for connections initiated from the proxy. This might happen when
the proxy does a health check. This is used in `lazymc` to poll the
server state and to prepare client connections.

See: https://www.haproxy.org/download/2.4/doc/proxy-protocol.txt